educa/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==2.2
django-braces==1.14.0
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-filter==2.4.0
django-memcache-status==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
python-memcached==1.59
pytz==2020.5
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.2
uWSGI==2.0.19.1
